requirements.txt
Flask
redis
python-dotenv
gunicorn
flask-cors